+++ corophage/src/prelude.rs
@@ -1,7 +1,8 @@
 pub use frunk::hlist;
 pub use frunk_core as frunk;
 
-pub use crate::{Cancelled, Control, Effect, Effectful, Effects, Never, Program, Yielder};
+#[allow(deprecated)]
+pub use crate::{Cancelled, Control, Eff, Effect, Effectful, Effects, Never, Program, Yielder};
 
 #[cfg(feature = "macros")]
 pub use crate::{effect, effectful};